requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.3.0=py39h06a4308_0
aiohttp=3.8.1=py39h7f8727e_1
aiosignal=1.2.0=pyhd3eb1b0_0
async-timeout=4.0.2=py39h06a4308_0
atk-1.0=2.36.0=ha1a6a79_0
attrs=22.1.0=py39h06a4308_0
blas=1.0=mkl
blinker=1.4=py39h06a4308_0
box2d-py=2.3.5=pypi_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py39h27cfd23_1003
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.10.11=h06a4308_0
cachetools=4.2.2=pyhd3eb1b0_0
cairo=1.16.0=h19f5f5c_2
certifi=2022.9.24=py39h06a4308_0
cffi=1.15.1=py39h5eee18b_2
charset-normalizer=2.1.1=pypi_0
click=8.0.4=py39h06a4308_0
cloudpickle=2.2.0=pypi_0
cryptography=38.0.1=py39h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd3eb1b0_0
dataclasses=0.8=pyh6d0b6a4_7
dbus=1.13.18=hb2f20db_0
deepdiff=6.2.1=pypi_0
dm-env=1.5=pypi_0
dm-env-rpc=1.1.3=pypi_0
dm-tree=0.1.7=pypi_0
enum34=1.1.10=pypi_0
expat=2.4.9=h6a678d5_0
fftw=3.3.9=h27cfd23_1
font-ttf-dejavu-sans-mono=2.37=0
font-ttf-inconsolata=2.000=0
font-ttf-source-code-pro=2.030=0
font-ttf-ubuntu=0.83=0
fontconfig=2.13.1=hef1e5e3_1
fonts-anaconda=1=h8fa9717_0
fonts-conda-ecosystem=1=hd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
fribidi=1.0.10=h7b6447c_0
frozenlist=1.2.0=py39h7f8727e_0
gdk-pixbuf=2.42.8=h433bba3_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=he621ea3_2
gobject-introspection=1.72.0=py39hbb6d50b_1
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
googleapis-common-protos=1.57.0=pypi_0
graphite2=1.3.14=h295c915_1
graphviz=2.50.0=h3cd0ef9_0
grpcio=1.50.0=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
gtk2=2.24.33=h73c1081_2
gts=0.7.6=hb67d8dd_3
gym=0.26.2=pypi_0
gym-notices=0.0.8=pypi_0
harfbuzz=4.3.0=hd55b92a_0
icu=58.2=he6710b0_3
idna=3.4=py39h06a4308_0
immutabledict=2.2.3=pypi_0
importlib-metadata=5.1.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jbig=2.1=0
joblib=1.1.1=py39h06a4308_0
jpeg=9e=h7f8727e_0
kiwisolver=1.4.2=py39h295c915_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libdeflate=1.8=h7f8727e_5
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgd=2.3.3=h695aa2c_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.14=0
libpng=1.6.37=hbc83047_0
libprotobuf=3.20.1=h4ff587b_0
librsvg=2.54.4=h19fe530_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.4.0=hecacb30_2
libtool=2.4.2=0
libuuid=1.0.3=0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxml2=2.9.14=h74e7548_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=py39h06a4308_0
matplotlib=3.5.3=py39h06a4308_0
matplotlib-base=3.5.3=py39hf590b9c_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
mock=4.0.3=pypi_0
mpyq=0.2.5=pypi_0
multidict=6.0.2=py39h5eee18b_0
munkres=1.1.4=py_0
ncurses=6.3=h5eee18b_3
ninja=1.7.2=0
numpy=1.23.4=py39h14f4228_0
numpy-base=1.23.4=py39h31eccc5_0
oauthlib=3.2.1=py39h06a4308_0
openssl=1.1.1s=h7f8727e_0
ordered-set=4.1.0=pypi_0
packaging=21.3=pyhd3eb1b0_0
pango=1.50.7=h05da053_0
pcre=8.45=h295c915_0
pillow=9.2.0=py39hace64e9_1
pip=22.2.2=py39h06a4308_0
pixman=0.40.0=h7f8727e_1
portpicker=1.5.2=pypi_0
protobuf=3.20.1=pypi_0
psutil=5.9.4=pypi_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pygame=2.1.0=pypi_0
pyjwt=2.4.0=py39h06a4308_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py39h06a4308_0
pyqt=5.9.2=py39h2531618_6
pysc2=4.0.0=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.15=h7a1cb2a_2
python-dateutil=2.8.2=pyhd3eb1b0_0
python-graphviz=0.20.1=pypi_0
pytorch=1.12.1=py3.9_cuda11.3_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
qt=5.9.7=h5867ecd_1
readline=8.2=h5eee18b_0
requests=2.28.1=py39h06a4308_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
s2clientprotocol=5.0.10.89165.0=pypi_0
s2protocol=5.0.10.88500.0=pypi_0
scikit-learn=1.1.3=py39h6a678d5_0
scipy=1.9.3=pypi_0
setuptools=65.5.0=py39h06a4308_0
sip=4.19.13=py39h295c915_0
six=1.16.0=pyhd3eb1b0_1
sk-video=1.1.10=pypi_0
sqlite=3.40.0=h5082296_0
swig=4.1.0=pypi_0
tensorboard=2.10.0=py39h06a4308_0
tensorboard-data-server=0.6.0=py39hca6d32c_0
tensorboard-plugin-wit=1.8.1=py39h06a4308_0
tensorboardx=2.2=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
tornado=6.2=py39h5eee18b_0
typing-extensions=4.4.0=pypi_0
typing_extensions=4.3.0=py39h06a4308_0
tzdata=2022f=h04d1e81_0
urllib3=1.26.13=pypi_0
websocket-client=1.4.2=pypi_0
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.6=h5eee18b_0
yarl=1.8.1=py39h5eee18b_0
zipp=3.11.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
